Abstract out some of the descriptor Span-parsing helpers
